+++ src/components/Storage/DesktopStorage/DesktopItemsList/Item/ListItem.tsx
@@ -14,7 +14,7 @@ import {
 } from '@chakra-ui/react'
 import { useState } from 'react'
 import { Item } from '../../../../../mongo/models/item'
-import ProductButton from '../../../../Overridden/ProductButton/ProductButton'
+import ProductButton from '../../../../UI/Custom Buttons/ProductButton/ProductButton'
 
 interface Props {
   item: Item
